sinai_nexus_backend/requirements.txt
numpy<2
fastapi
uvicorn
python-dotenv
pandas
pyarrow
rapidfuzz
google-generativeai
python-multipart
sentence-transformers
scikit-learn
pdfplumber>=0.11.8
unstructured[docx]
supabase